requirements.txt
fastapi
uvicorn
pymongo
pydantic
openai
python-dotenv

